fix(gki-5.10): mrdump kallsyms_markers/num_syms 类型改回 unsigned int
scripts/kallsyms.c 以 .long(4B) 生成 kallsyms_num_syms(L455)/kallsyms_markers(L484)，kernel/kallsyms.c(45/54) 与 debug_kinfo.c(31/40) 均声明 unsigned int；mrdump 原写 unsigned long(arm64 8B) 错位。mrdump 对两符号仅取地址(&sym/数组退化)，从不按元素读值，故改宽不影响链接/地址，纯对齐 5.10 真相。

+++ drivers/misc/mediatek/aee/mrdump/mrdump_private.h
@@ -44,8 +44,8 @@ extern const int kallsyms_offsets[] __weak;
 extern const u8 kallsyms_names[] __weak;
 extern const u8 kallsyms_token_table[] __weak;
 extern const u16 kallsyms_token_index[] __weak;
-extern const unsigned long kallsyms_markers[] __weak;
-extern const unsigned long kallsyms_num_syms
+extern const unsigned int kallsyms_markers[] __weak;
+extern const unsigned int kallsyms_num_syms
 __attribute__((weak, section(".rodata")));
 
 #ifdef MODULE
